refactor(driver): move querylangdriver to a mixin class
