+++ jina/drivers/querylang/reverse.py
@@ -3,14 +3,13 @@ __license__ = "Apache-2.0"
 
 from typing import Iterable
 
-from . import QueryLangDriver
+from .. import QuerysetReader, BaseRecursiveDriver
 
 if False:
     from ...proto import jina_pb2
 
 
-class ReverseQL(QueryLangDriver):
-
+class ReverseQL(QuerysetReader, BaseRecursiveDriver):
     """Reverses the order of the provided ``docs``.
 
         This is often useful when the proceeding Pods require only a signal, not the full message.
